consolidated method for validating required documents to PipelineMain
