+++ src/main/java/edu/cuanschutz/ccp/tm_provider/etl/fn/FilterUnactionableTextFn.java
@@ -22,6 +22,7 @@ import edu.cuanschutz.ccp.tm_provider.etl.PipelineMain;
 import edu.cuanschutz.ccp.tm_provider.etl.ProcessingStatus;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DocumentCriteria;
 import edu.cuanschutz.ccp.tm_provider.etl.util.DocumentType;
+import edu.cuanschutz.ccp.tm_provider.etl.util.PipelineKey;
 import edu.ucdenver.ccp.common.collections.CollectionsUtil;
 import edu.ucdenver.ccp.nlp.core.annotation.TextAnnotation;
 import lombok.Data;
@@ -56,7 +57,7 @@ public class FilterUnactionableTextFn extends DoFn<KV<String, String>, KV<String
 	public static PCollectionTuple process(
 			PCollection<KV<ProcessingStatus, Map<DocumentCriteria, String>>> statusEntity2Content,
 			DocumentCriteria outputDocCriteria, Set<DocumentCriteria> requiredDocumentCriteria,
-			com.google.cloud.Timestamp timestamp) {
+			com.google.cloud.Timestamp timestamp, PipelineKey pipelineKey) {
 
 		return statusEntity2Content.apply("filter unactionable text", ParDo.of(
 				new DoFn<KV<ProcessingStatus, Map<DocumentCriteria, String>>, KV<ProcessingStatus, List<String>>>() {
@@ -71,13 +72,16 @@ public class FilterUnactionableTextFn extends DoFn<KV<String, String>, KV<String
 						try {
 							// check to see if all documents are present
 							Map<DocumentCriteria, String> docs = statusEntityToText.getValue();
-							if (!docs.keySet().equals(requiredDocumentCriteria)) {
-								PipelineMain.logFailure(ETL_FAILURE_TAG,
-										"Unable to complete post-processing due to missing annotation documents for: "
-												+ docId + " -- contains (" + docs.size() + ") "
-												+ docs.keySet().toString(),
-										outputDocCriteria, timestamp, out, docId, null);
-							} else {
+							if (PipelineMain.requiredDocumentsArePresent(docs.keySet(), requiredDocumentCriteria,
+									pipelineKey, ETL_FAILURE_TAG, docId, outputDocCriteria, timestamp, out)) {
+//							if (!PipelineMain.fulfillsRequiredDocumentCriteria(docs.keySet(),
+//									requiredDocumentCriteria)) {
+//								PipelineMain.logFailure(ETL_FAILURE_TAG,
+//										"Unable to complete filter unactionable text due to missing annotation documents for: "
+//												+ docId + " -- contains (" + docs.size() + ") "
+//												+ docs.keySet().toString(),
+//										outputDocCriteria, timestamp, out, docId, null);
+//							} else {
 
 								Map<DocumentType, Collection<TextAnnotation>> docTypeToContentMap = PipelineMain
 										.getDocTypeToContentMap(statusEntity.getDocumentId(), docs);
