+++ src/test/java/edu/cuanschutz/ccp/tm_provider/etl/PipelineMainTest.java
@@ -7,7 +7,6 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -656,6 +655,48 @@ public class PipelineMainTest {
 	@Test(expected = java.lang.NumberFormatException.class)
 	public void testGetSemanticVersion_e2() {
 		assertArrayEquals(new int[] { 1, 3, 4 }, PipelineMain.getSemanticVersion(".3.4"));
+	}
+
+	@Test
+	public void testFulfillsRequiredDocumentCriteria() {
+
+		Set<DocumentCriteria> requiredDocumentCriteria = new HashSet<DocumentCriteria>();
+
+		DocumentCriteria dc1 = new DocumentCriteria(DocumentType.CRF_CRAFT, DocumentFormat.BIONLP, PipelineKey.CRF,
+				PipelineMain.PIPELINE_VERSION_RECENT);
+		DocumentCriteria dc2 = new DocumentCriteria(DocumentType.TEXT, DocumentFormat.TEXT,
+				PipelineKey.MEDLINE_XML_TO_TEXT, "0.1.1");
+		DocumentCriteria dc3 = new DocumentCriteria(DocumentType.ABBREVIATIONS, DocumentFormat.BIONLP,
+				PipelineKey.ABBREVIATION, PipelineMain.PIPELINE_VERSION_RECENT);
+
+		requiredDocumentCriteria.add(dc1);
+		requiredDocumentCriteria.add(dc2);
+		requiredDocumentCriteria.add(dc3);
+
+		Set<DocumentCriteria> docCriteria = new HashSet<DocumentCriteria>();
+
+		DocumentCriteria dc1a = new DocumentCriteria(DocumentType.CRF_CRAFT, DocumentFormat.BIONLP, PipelineKey.CRF,
+				"1.1.1");
+		DocumentCriteria dc2a = new DocumentCriteria(DocumentType.TEXT, DocumentFormat.TEXT,
+				PipelineKey.MEDLINE_XML_TO_TEXT, "0.1.1");
+		DocumentCriteria dc3a = new DocumentCriteria(DocumentType.ABBREVIATIONS, DocumentFormat.BIONLP,
+				PipelineKey.ABBREVIATION, "0.3.0");
+		docCriteria.add(dc1a);
+		docCriteria.add(dc2a);
+		docCriteria.add(dc3a);
+
+		assertTrue(PipelineMain.fulfillsRequiredDocumentCriteria(docCriteria, requiredDocumentCriteria));
+
+		docCriteria = new HashSet<DocumentCriteria>();
+
+		DocumentCriteria dc2b = new DocumentCriteria(DocumentType.TEXT, DocumentFormat.TEXT,
+				PipelineKey.MEDLINE_XML_TO_TEXT, "0.1.2");
+		docCriteria.add(dc1a);
+		docCriteria.add(dc2b); // version mismatch
+		docCriteria.add(dc3a);
+
+		assertFalse(PipelineMain.fulfillsRequiredDocumentCriteria(docCriteria, requiredDocumentCriteria));
+
 	}
 
 }
